Short-circuit sidebar poll filter conditional
When there are exceptions elsewhere in the website, this template tag often gets called anyway, with `user` set to `None`. This causes an `AttributeError` when checking `is_authenticated`, which spams the log "fake" error messages that hides the true error message. Checking if `user` exist firsts works around this issue.