Drop `BootstrapTableExtension` in favour of Bleach filter
This is much more universal, catching all elements, not just those created early in ther Markdown filter. This allows using the filter for `img` elements, which means we can drop the `img-fluid` replacemen in `base.scss`.

`text-break` is now applied to anchors with this new filter, instead of
through a callback in `LinkifyFilter`. This is definitely a debatable
change, but I prefer not setting classes like this in two seperate
places.