Fix dashboard comment truncation
Comment truncation on the dashboard broke two months ago, when we implemented our inline markdown filter (!340). We were previously relying on the comment being implicitly turned into a string to get its (truncated) comment text, but the filters require a string input and does not perform a conversion on its own. Because of this, the dashboard was changed to explicitly display a comment's text, which doesn't truncate that text (since code for that used to live in the `__str__` method, exclusively.) This issue has been rectified by moving truncation functionality into a new `truncated` method.