Show an unlimited number of autocomplete options
I've received feedback that some people don't understand that you can search in an autocomplete widget. Instead, they try to scroll down to the option they want to select, get stuck after 50 entries, and then give up. This commit partially solves this issue by showing an unlimited number of options. The downside is potentially performance, but I don't think we ever have enough entries for this to be a problem (it isn't a problem on my PC, at least, though it is a comparatively high-end device)